pyproject.toml
[build-system]
requires = ["setuptools>=61.0",
            "versioningit~=2.0",]

build-backend = "setuptools.build_meta"

# Self-descriptive entries which should always be present
# https://packaging.python.org/en/latest/specifications/declaring-project-metadata/
[project]
name = "idptools-starling"
description = "Construction of intrinsically disordered proteins ensembles through multiscale generative models"
dynamic = ["version"]
readme = "README.md"
authors = [
    { name = "Borna Novak", email = "bnovak@wustl.edu" }
]
license = { text = "MIT" }
# See https://pypi.org/classifiers/
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
]
requires-python = ">=3.8"
# Declare any run-time dependencies that should be installed with the package.
dependencies = [
	'numpy>=1.26',
	'torch',
	'scipy',
	'cython>=3.0.0',
	'matplotlib',
	'jupyter',
	'ipython',
	'pytorch-lightning',
	'scikit-learn',
	'einops',
	'tqdm',
	'PyYAML',
	'h5py',
	'pandas',
	'pytest',
	'protfasta',
	'soursop',
	'hdf5plugin',
	'mdtraj>=1.9.7',
	'metapredict>=3.0',
	'faiss-cpu>=1.7.4',
	'zstandard>=0.22',
    ]
	#'finches @ git+https://git@github.com/idptools/finches@main#egg=finches',
	#'sparrow @ git+https://git@github.com/idptools/sparrow.git@main#egg=sparrow.git']

# Update the urls once the hosting is set up.
#[project.urls]
#"Source" = "https://github.com/<username>/starling/"
#"Documentation" = "https://starling.readthedocs.io/"

[project.scripts]
starling-vae-train = "starling.training.vae_train:train_vae"
starling-ddpm-train = "starling.training.diffusion_train:train_model"
starling-sample = "starling.training.vae_generate:vae_generate"
ae-train = "starling.training.ae_train:train_ae"

starling = "starling.scripts.starling_main_cli:main"
starling-benchmark = "starling.scripts.starling_main_cli:starling_benchmark"

starling2xtc = "starling.scripts.starling_converter:starling2xtc"
starling2pdb = "starling.scripts.starling_converter:starling2pdb"
starling2numpy = "starling.scripts.starling_converter:starling2numpy"
starling2info = "starling.scripts.starling_converter:starling2info"
starling2sequence = "starling.scripts.starling_converter:starling2sequence"
starling2starling = "starling.scripts.starling_converter:starling2starling"
numpy2starling = "starling.scripts.starling_converter:numpy2starling"
xtc2starling = "starling.scripts.starling_converter:xtc2starling"
starling-search = "starling.scripts.starling_search:main"
starling-pretokenize = "starling.scripts.starling_pretokenize:main"


[project.optional-dependencies]
search-gpu = [
  "faiss-gpu>=1.7.4"
]
test = [
  "pytest>=6.1.2",
  "pytest-runner"
]

[tool.setuptools]
# This subkey is a beta stage development and keys may change in the future, see https://setuptools.pypa.io/en/latest/userguide/pyproject_config.html for more details
#
# As of version 0.971, mypy does not support type checking of installed zipped
# packages (because it does not actually import the Python packages).
# We declare the package not-zip-safe so that our type hints are also available
# when checking client code that uses our (installed) package.
# Ref:
# https://mypy.readthedocs.io/en/stable/installed_packages.html?highlight=zip#using-installed-packages-with-mypy-pep-561
zip-safe = false
# Let setuptools discover the package in the current directory,
# but be explicit about non-Python files.
# See also:
#   https://setuptools.pypa.io/en/latest/userguide/pyproject_config.html#setuptools-specific-configuration
# Note that behavior is currently evolving with respect to how to interpret the
# "data" and "tests" subdirectories. As of setuptools 63, both are automatically
# included if namespaces is true (default), even if the package is named explicitly
# (instead of using 'find'). With 'find', the 'tests' subpackage is discovered
# recursively because of its __init__.py file, but the data subdirectory is excluded
# with include-package-data = false and namespaces = false.
include-package-data = false
[tool.setuptools.packages.find]
namespaces = false
where = ["."]

# Ref https://setuptools.pypa.io/en/latest/userguide/datafiles.html#package-data
[tool.setuptools.package-data]
starling = [
    "py.typed",
	"configs/**"
]

[tool.versioningit]
default-version = "1+unknown"

[tool.versioningit.format]
distance = "{base_version}+{distance}.{vcs}{rev}"
dirty = "{base_version}+{distance}.{vcs}{rev}.dirty"
distance-dirty = "{base_version}+{distance}.{vcs}{rev}.dirty"

[tool.versioningit.vcs]
# The method key:
method = "git"  # <- The method name
# Parameters to pass to the method:
match = ["*"]
default-tag = "1.0.0"

[tool.versioningit.write]
file = "starling/_version.py"
